refactor(moveRule): movement 검증 클래스명 변경
